Unify reponav-dropdown creation
